requirements.txt
torch==2.5.1
torchvision==0.20.1
opencv-python==4.11.0.86
diffusers==0.32.2
transformers==4.52.4
tokenizers==0.21.0
accelerate==1.7.0
gradio==5.31.0
numpy==1.26.4
llamafactory==0.9.3
tqdm
matplotlib
imageio
ffmpeg-python
easydict
ftfy
dashscope
imageio-ffmpeg
decord
einops
scikit-image
scikit-learn
pycocotools
timm
onnxruntime-gpu
BeautifulSoup4